bq_fsp/pyproject.toml
[build-system]
requires = ["setuptools>=67", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "bq-fsp"
version = "0.1.0"
description = "BLUX Quantum: Find Suspicious Patterns"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "typer>=0.12",
    "rich>=13.7",
    "pathspec>=0.12",
    "regex>=2024.5",
    "ruamel.yaml>=0.18"
]

[project.optional-dependencies]
dev = ["pytest", "black", "mypy"]

[project.scripts]
"bq-fsp" = "bq_fsp.cli:app"

[tool.bq.plugin]
entry = "bq_fsp.cli:register"
